Move the remoteca controller to its own pkg to deal with an import cycle
